Updates key-verification logic for dual keys
The CI logic explicitly checks that the prod key was used to sign RPMs
in this repo. We're currently rotating prod signing keys, so I've added
supported for *both* keys to the script.

This change doesn't include strict logic about all newer RPMs being
signed with the newer key, so humans should still verify the correct key
was used, as we get closer to the expiry date. It'll still catch
easy mistakes like if the test key is used, since we frequently promote
the same debs from yum-test to yum prod.