Resigns sdw-dom0-config 0.5.5 with new key
As part of the 2021 key rotation, we'll simply re-sign the existing RPM
for v0.5.5 with the new key, to simplify the first run story.
In the near future, we'll release a new version and make sure that's
signed, but until that's done, we can still ensure that the prod repo's
most recent versions are all properly signed with the new key.