Add script to commit repodata/
Let's match the securedrop-apt-prod process by generating metadata at
commit-time instead of doing it on the server.

The publish script takes care to generate reproducible output by
fixing the mtime of all the RPMs and telling `createrepo_c` what the
time should be.

CI verifies the generated metadata is up to date and fully reproducible
using the `--reproduce` flag.