Simplify CI verification script, move to GitHub Actions
It only takes a few seconds to run, so let's verify everything, all the
time. The code is simpler now, we let exceptions bubble up instead of
catching them and re-printing the error message.

We also don't need to bundle the PGP keys in this repo, we can download
them from securedrop-builder at runtime (just like how we do it in
securedrop-apt-prod).